rename (virtual) wardrobe to dressme
