+++ README.md
@@ -1,3 +1,3 @@
-# DressMe
+# dressme
 
 A virtual wardrobe that shows you how clothes look on you.
