+++ api/main.py
@@ -25,10 +25,10 @@ from replicate.client import Client
 from sqlalchemy.orm import joinedload
 from sqlmodel import Session, select
 
-from .wardrobe import db
-from .wardrobe.auth import verify_token
-from .wardrobe.combining import combine_wearables
-from .wardrobe.settings import get_settings
+from .dressme import db
+from .dressme.auth import verify_token
+from .dressme.combining import combine_wearables
+from .dressme.settings import get_settings
 
 settings = get_settings()
 replicate = Client(api_token=settings.REPLICATE_API_TOKEN)
